Add initial data structure implementation
Based on the idea of a hash table, with multiple levels of bins for date
parts. Though no need to use an actual hash as the data is already stored
as ints.

I've put this commit in as a working checkpoint. But there are a few
parts to revisit and maybe refactor. Namely merging the arrays when returning
dates.