Change DateRecorder to use dynamic YearBins
The initial implementation worked, but it was wasteful. Although the
memory footprint wasn't large in comparison to some programs, it was
overkill to have so many empty bins. Especially if the analyser
were to be used on a narrow set of logs within one year.

It also couldn't handle any logs with dates after the year the code
was written. So I thought it would be better to have a dynamic size.
This has increased the code complexity for the Add method, but is a
more elegant way to handle the years.